requirements.txt
# RAG+LLM表格生成项目依赖

# 核心依赖
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0

# 配置和工具
PyYAML>=6.0
tqdm>=4.64.0

# 可选依赖（用于更好的性能）
accelerate>=0.20.0
bitsandbytes>=0.39.0

# 开发和调试
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 注意：
# 1. 如果使用CUDA，请确保安装了对应版本的PyTorch CUDA版本
# 2. 如果遇到模型下载问题，可能需要配置Hugging Face镜像源
# 3. 建议使用Python 3.8+版本
